Fix warnings
* add missing return statement
* fix format control strings
  - use OMR_PRI* macros from omrformatconsts.h
* fix casts of int to pointer type of differing size
* cast to 64-bits before doing 32-bit shift
* use 0 instead of NULL for non-pointer types
* remove redundant casts
* fix indentation (use spaces instead of tabs)

+++ compiler/aarch64/codegen/OMRTreeEvaluator.cpp
@@ -2495,7 +2495,7 @@ addMetaDataForLoadAddressConstantFixed(TR::CodeGenerator *cg, TR::Node *node, TR
          relo = new (cg->trHeapMemory()) TR::BeforeBinaryEncodingExternalRelocation(
             firstInstruction,
             (uint8_t *)node->getSymbolReference(),
-            (uint8_t *)node->getInlinedSiteIndex(),
+            (uint8_t *)(intptr_t)node->getInlinedSiteIndex(),
             TR_DataAddress, cg);
          break;
          }
@@ -2529,7 +2529,7 @@ addMetaDataForLoadAddressConstantFixed(TR::CodeGenerator *cg, TR::Node *node, TR
             relo = new (cg->trHeapMemory()) TR::BeforeBinaryEncodingExternalRelocation(
                firstInstruction,
                (uint8_t *)symRef,
-               (uint8_t *)(node == NULL ? -1 : node->getInlinedSiteIndex()),
+               (uint8_t *)(intptr_t)(node == NULL ? -1 : node->getInlinedSiteIndex()),
                TR_ClassAddress, cg);
             }
          break;
